separated out IO registers

+++ GBAEmulator/Visual.cs
@@ -241,7 +241,8 @@ namespace GBAEmulator
                 this.FramesUntilSaveDump = SAVE_DELAY;
             }
             
-            this.Text = string.Format("GBAC-  : {0} <{1:0.0} fps>", this.gba.mem.ROMName, (1000 * this.gba.ppu.frame / (ulong)this.FPSTimer.ElapsedMilliseconds));
+            this.Text = string.Format("GBAC-  : {0} <{1:0.0} fps>",
+                this.gba.mem.ROMName, (1000 * this.gba.ppu.frame / (double)this.FPSTimer.ElapsedMilliseconds));
 
             if (this.FPSTimer.ElapsedMilliseconds > 2000)
             {
